Added window confirm at delete project

+++ src/actions/projectActions.js
@@ -44,12 +44,14 @@ export const getProject = (id, history) => async dispatch => {
 
 };
 export const deleteProject = id => async dispatch => {
+        if(window.confirm("Are you sure? This will delete the project and all the data related to it")){
+            const res = await axios.delete(`/api/project/${id}` );
+            dispatch({
+                type:DELETE_PROJECT,
+                payload: id
+            });
+        }
 
-        const res = await axios.delete(`/api/project/${id}` );
-        dispatch({
-            type:DELETE_PROJECT,
-            payload: id
-        });
 
 
 };
